requirements.txt
amqp==5.1.1
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
django-celery-results==2.5.1
djangorestframework==3.14.0
idna==3.4
kombu==5.3.7
prompt-toolkit==3.0.39
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.6
vine==5.1.0
wcwidth==0.2.8
